refactor: consolidate and optimize async storage utility functions
