+++ src/persistentStorage/asyncStorageShared.ts
@@ -1,6 +1,9 @@
 import { rc_parse_json, rc_string, rc_tuple } from 'runcheck';
 import type {
+  AsyncStorageEntryMetadata,
+  AsyncStorageMetadataOrder,
   AsyncStorageNamespaceScope,
+  AsyncStorageNamespaceStaticPolicy,
   AsyncStorageProtectedEntryRef,
 } from './types';
 import { parseAsyncStorageNamespaceKind } from './types';
@@ -51,4 +54,49 @@ export function parseAsyncStorageRecordKey(
   }
 
   return { rawKey: key, recordKind: 'raw' };
+}
+
+export function normalizeStaticPolicy(
+  policy: AsyncStorageNamespaceStaticPolicy | null | undefined,
+): AsyncStorageNamespaceStaticPolicy | null {
+  if (policy === undefined || policy === null) return null;
+
+  const maxEntries =
+    typeof policy.maxEntries === 'number' &&
+    Number.isInteger(policy.maxEntries) &&
+    policy.maxEntries >= 0
+      ? policy.maxEntries
+      : undefined;
+  const pinnedKeys = Array.isArray(policy.pinnedKeys)
+    ? [...new Set(policy.pinnedKeys)].sort((left, right) =>
+        left.localeCompare(right),
+      )
+    : [];
+
+  if (maxEntries === undefined && pinnedKeys.length === 0) {
+    return null;
+  }
+
+  return {
+    ...(maxEntries !== undefined ? { maxEntries } : {}),
+    ...(pinnedKeys.length > 0 ? { pinnedKeys } : {}),
+  };
+}
+
+export function compareMetadata(
+  left: AsyncStorageEntryMetadata<Record<string, unknown>>,
+  right: AsyncStorageEntryMetadata<Record<string, unknown>>,
+  order: AsyncStorageMetadataOrder,
+): number {
+  if (order === 'key') {
+    return left.key.localeCompare(right.key);
+  }
+
+  if (left.lastAccessAt !== right.lastAccessAt) {
+    return order === 'lru-asc'
+      ? left.lastAccessAt - right.lastAccessAt
+      : right.lastAccessAt - left.lastAccessAt;
+  }
+
+  return left.key.localeCompare(right.key);
 }
