+++ src/persistentStorage/indexedDbAsyncStorageAdapter.ts
@@ -2,8 +2,10 @@ import { __LEGIT_CAST__ } from '@ls-stack/utils/saferTyping';
 import { createAsyncStorageAdapter } from './asyncStorageAdapter';
 import {
   ASYNC_NAMESPACE_INDEX_RECORD_KEY,
+  compareMetadata,
   getNamespaceId,
   getPayloadRecordKey,
+  normalizeStaticPolicy,
   parseAsyncStorageRecordKey,
 } from './asyncStorageShared';
 import type {
@@ -307,33 +309,6 @@ function toPublicMetadata(
   };
 }
 
-function normalizeStaticPolicy(
-  policy: AsyncStorageNamespaceStaticPolicy | null | undefined,
-): AsyncStorageNamespaceStaticPolicy | null {
-  if (policy === undefined || policy === null) return null;
-
-  const maxEntries =
-    typeof policy.maxEntries === 'number' &&
-    Number.isInteger(policy.maxEntries) &&
-    policy.maxEntries >= 0
-      ? policy.maxEntries
-      : undefined;
-  const pinnedKeys = Array.isArray(policy.pinnedKeys)
-    ? [...new Set(policy.pinnedKeys)].sort((left, right) =>
-        left.localeCompare(right),
-      )
-    : [];
-
-  if (maxEntries === undefined && pinnedKeys.length === 0) {
-    return null;
-  }
-
-  return {
-    ...(maxEntries !== undefined ? { maxEntries } : {}),
-    ...(pinnedKeys.length > 0 ? { pinnedKeys } : {}),
-  };
-}
-
 const INDEXED_DB_ENTRY_RECORD_KEYS = new Set([
   'a',
   'd',
@@ -458,24 +433,6 @@ function expandEntryValue(
 ): unknown {
   void scope;
   return record.d;
-}
-
-function compareMetadata(
-  left: AsyncStorageEntryMetadata<Record<string, unknown>>,
-  right: AsyncStorageEntryMetadata<Record<string, unknown>>,
-  order: AsyncStorageMetadataOrder,
-): number {
-  if (order === 'key') {
-    return left.key.localeCompare(right.key);
-  }
-
-  if (left.lastAccessAt !== right.lastAccessAt) {
-    return order === 'lru-asc'
-      ? left.lastAccessAt - right.lastAccessAt
-      : right.lastAccessAt - left.lastAccessAt;
-  }
-
-  return left.key.localeCompare(right.key);
 }
 
 function openRequestAsPromise<T>(request: IDBRequest): Promise<T> {
